Add Starry test harness workflow (#15)
* Title: Add Starry test harness workflowMessage:- add starry-test-harness CI workflow triggered on push/pull_request- mirror check/test concurrency settings to avoid duplicate runs- run make ci-test run via harness repo and upload suite logs/artifacts

* Fix clippy errors and improve Qemu test

* ci: install musl + warm up test runner, add rust targets for check

* ci: set vsock in test

* ci: set vsock in citest

* chore(arceos): sync doc_cfg fixes

* fix: silence new clippy warnings

* chore(arceos): pull latest doc_cfg fixes

* chore(arceos): bump to latest main

---------

+++ api/src/terminal/ldisc.rs
@@ -349,7 +349,7 @@ impl<R: TtyRead, W: TtyWrite> LineDiscipline<R, W> {
             term.special_char(VMIN) as usize
         };
 
-        if buf.len() < vmin as usize {
+        if buf.len() < vmin {
             return Err(AxError::WouldBlock);
         }
